Remove import from _custom_element_types in test (#136)
Remove import from _custom_element_types in test because the module is
removed in the latest version of onnx. I chose to remove the entries
instead of doing version checks because I don't expect the logic to be
broken for older onnx versions. When we bump the lowest supported onnx
version to be 1.19, we can remove the logic handling the custom types in
the from_numpy() method as well.

---------

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>