Fix support for torch.float8_e8m0fnu for older versions of PyTorch (#145)
`torch.float8_e8m0fnu` was added only in PyTorch 2.7. In the
`to_torch_dtype` method where we build a dictionary of type mapping from
onnx dtypes to torch dtypes, a undefinied member error would be raised
if user has a torch version <2.7.

This change conditionally adds the float8_e8m0fnu to the mapping if it
is definied in torch. When users request the dtype in an environment
that has the older version of PyTorch, an error will be raised to
suggest users to upgrade torch version.

---------

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>