Support string tensors in DeduplicateInitializersPass and call_onnx_api
Getting the number of bytes in a string tensor needs special treatment
as the STRING data type does not define a bitwidth and the size needs to
be computed from flattening the strings into a sequence of bytes. See
call_onnx_api querying .nbytes to temporarily remove large initializers.

String initializers need to be converted to bytes via the .string_data
method instead of the .tobytes method in DeduplicateInitializersPass.

Signed-off-by: Christoph Berganski <christoph.berganski@gmail.com>