Support string tensors in DeduplicateInitializersPass and call_onnx_api (#169)
Getting the number of bytes in a string tensor needs special treatment
as the STRING data type does not define a bitwidth but needs to be
computed from flattening the strings into a sequence of bytes. See
call_onnx_api querying .nbytes to temporarily remove large initializers.

String initializers need to be converted to bytes via the .string_data
method instead of the .tobytes method in DeduplicateInitializersPass.

Also adds mapping of `np.bytes_` to `STRING` in `DataType.from_numpy`,
which was probably just overlooked before, as object and `np.str_`
(unicode strings) are already handled. This is related to
https://github.com/microsoft/onnxscript/pull/2514

---------

Signed-off-by: Christoph Berganski <christoph.berganski@gmail.com>