Simplify attribute signatures and use tuples for repeating attrs (#186)
Always return a tuple for repeated attributes so that users do not need
to convert the type again when comparing values.

Note that previously the methods are annotated to return Sequence, and
they return list as a concreate type. This PR updates the methods to
return tuples, which may be a BC breaking behavior, but should be good
in the long run to reflect the immutable nature of the attributes.

---------

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>