Implement replace_all_uses_with on Value (#235)
Given it is a fundamental operation in graph transformation, I added it
as a method on `Value` so that it is more convenient and discoverable.
The ir.convenience.replace_all_uses_with function is preserved to handle
batch processing cases on multiple values.

Also added a `replace_graph_outputs` option to allow users to replace
graph outputs as well. Users are still responsible to assign the output
name to the replacement value if they want to preserve the signature.

### BC breaking

An error is raised when `replace_graph_outputs` is False && when the
value to replace is a graph output.

---------

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>