Minor cleanup of cloning logic in inliner (#252)
Minor changes to the cloning logic used in the inliner. This cleanup is
first step to a couple of potential features, namely (a) Providing more
control to users on renaming strategy (when a value called X is inlined
at a callsite), and/or (b) Exposing the cloner as a standalone utility
(for users who want to implement some other variation).

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>